services/py-worker/requirements.runtime.txt
boto3==1.37.11
botocore==1.37.11
psycopg[binary]==3.2.6
pypdf==5.4.0
python-docx==1.1.2
qdrant-client==1.13.3
redis==5.2.1
httpx==0.28.1
python-json-logger==2.0.7
jieba==0.42.1
